refactor(reports): split html and table builders behind facade
